FEAT : alert #23 ask for confirmation before critical user action
초대 수락&거절, 멤버 탈퇴&초대 취소, 그룹 삭제&탈퇴 전에 confirm 창 띄움

+++ src/views/GroupInfo.tsx
@@ -21,8 +21,10 @@ const GroupInfo = () => {
 
     async function handleClickDelete() {
         if (params.teamId) {
-            await handleDeleteTeam(Number(params.teamId));
-            navigate('/myschedule');
+            if (confirm(`그룹을 삭제할까요?`)) {
+                await handleDeleteTeam(Number(params.teamId));
+                navigate('/myschedule');
+            }
         }
     }
